pyproject.toml
[project]
name = "languagebench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "evaluate>=0.4.3",
    "joblib>=1.4.2",
    "openai>=1.52.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "sacrebleu>=2.4.3",
    "tqdm>=4.66.6",
]
